split filter_factory.py off from filterbroker Signed-off-by: Chipmuenk <mail@chipmuenk.de>
